requirements.txt
googlesearch-python==1.0.0
PyYAML==5.3.1
pytest==6.2.1
requests==2.24.0
pandas==1.1.4
fastparquet==0.4.1
unidecode==1.2.0
vcrpy=4.1.1
future==0.18.2
fake-useragent==0.1.11
beautifulsoup4==4.9.1
selenium==2.53.6